requirements.txt
# ── LangChain + LangGraph ──────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0

# ── LLM provider ──────────────────────────────────────────────────────────
groq>=0.11.0

# ── Vector DB + Embeddings ────────────────────────────────────────────────
pinecone-client>=3.0.0
langchain-pinecone>=0.1.0
sentence-transformers>=3.0.0

# ── FastAPI ───────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ── Utils ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pypdf>=4.2.0
pydantic>=2.7.0
tenacity>=8.3.0
aiosmtplib>=3.0.0
jinja2>=3.1.4
httpx>=0.27.0
aiofiles>=23.2.1
